requirements.txt
# Core dependencies
requests>=2.28.0
argparse>=1.4.0

# LLM Provider APIs
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.1.0

# Data processing
numpy>=1.20.0
pandas>=1.3.0

# Utilities
tqdm>=4.64.0
